Render interface: minor code cleanups
* consistency: interface method `DestroyHW()` renamed to `Destroy()`
* `D3DXRenderBase`: not related to scene graph methods moved into separate file
* cleanup: redundant `virtual` specificators removed
