+++ src/Layers/xrRenderPC_R1/xrRender_R1.vcxproj.filters
@@ -959,6 +959,9 @@
     <ClCompile Include="FStaticRender_Shaders.cpp">
       <Filter>Core</Filter>
     </ClCompile>
+    <ClCompile Include="..\xrRender\D3DXRenderBase.cpp">
+      <Filter>Core</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <Text Include="Pipeline.txt" />
