+++ src/Layers/xrRenderPC_R2/xrRender_R2.vcxproj.filters
@@ -1103,6 +1103,9 @@
     <ClCompile Include="r2_shaders.cpp">
       <Filter>Core</Filter>
     </ClCompile>
+    <ClCompile Include="..\xrRender\D3DXRenderBase.cpp">
+      <Filter>Core</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <Text Include="todo.txt" />
