+++ src/Layers/xrRenderPC_R3/xrRender_R3.vcxproj.filters
@@ -1277,6 +1277,9 @@
     <ClCompile Include="r3_shaders.cpp">
       <Filter>Core</Filter>
     </ClCompile>
+    <ClCompile Include="..\xrRender\D3DXRenderBase.cpp">
+      <Filter>Core</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <Natvis Include="..\..\TypeHelper.natvis" />
