+++ src/Layers/xrRenderPC_R4/xrRender_R4.vcxproj.filters
@@ -1313,6 +1313,9 @@
     <ClCompile Include="r4_shaders.cpp">
       <Filter>Core</Filter>
     </ClCompile>
+    <ClCompile Include="..\xrRender\D3DXRenderBase.cpp">
+      <Filter>Core</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <Natvis Include="..\..\TypeHelper.natvis" />
