+++ src/xrEngine/Render.h
@@ -397,7 +397,7 @@ public:
     //	Destroy
     virtual void OnDeviceDestroy(bool bKeepTextures) = 0;
     virtual void ValidateHW() = 0;
-    virtual void DestroyHW() = 0;
+    virtual void Destroy() = 0;
     virtual void Reset(SDL_Window* hWnd, u32& dwWidth, u32& dwHeight, float& fWidth_2, float& fHeight_2) = 0;
     //	Init
     virtual void SetupStates() = 0;
